requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
crewai[litellm]==0.102.0
langchain-groq==0.1.3
python-dotenv==1.0.1
pydantic==2.11.10
streamlit==1.32.0
requests==2.31.0